// src/_proto/Protos/bookmarks.ts
export interface GetBookmarksRequest {
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.ReadingListType listType = 1;
     */
    listType: ReadingListType;
    /**
     * @generated from protobuf oneof: selector
     */
    selector: {
        oneofKind: "novelId";
        /**
         * @generated from protobuf field: int32 novelId = 2;
         */
        novelId: number;
    } | {
        oneofKind: "novelSlug";
        /**
         * @generated from protobuf field: string novelSlug = 3;
         */
        novelSlug: string;
    } | {
        oneofKind: undefined;
    };
    /**
     * @generated from protobuf oneof: pagination
     */
    pagination: {
        oneofKind: "pageInfo";
        /**
         * @generated from protobuf field: wuxiaworld.api.v2.PageInfoRequest pageInfo = 4;
         */
        pageInfo: PageInfoRequest;
    } | {
        oneofKind: "cursor";
        /**
         * @generated from protobuf field: wuxiaworld.api.v2.GetBookmarksRequest.CursorRequest cursor = 5;
         */
        cursor: GetBookmarksRequest_CursorRequest;
    } | {
        oneofKind: undefined;
    };
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.GetBookmarksRequest.SortBy sortBy = 6;
     */
    sortBy: GetBookmarksRequest_SortBy;
    /**
     * @generated from protobuf field: wuxiaworld.api.v2.GetBookmarksRequest.SortDirection sortDirection = 7;
     */
    sortDirection: GetBookmarksRequest_SortDirection;
}
/**
 * @generated from protobuf message wuxiaworld.api.v2.GetBookmarksRequest.CursorRequest
 */
export interface GetBookmarksRequest_CursorRequest {
    /**
     * @generated from protobuf field: int32 chapterOffset = 1;
     */
    chapterOffset: number;
    /**
     * @generated from protobuf field: int32 count = 2;
     */
    count: number;
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.GetBookmarksRequest.SortBy
 */
export enum GetBookmarksRequest_SortBy {
    /**
     * @generated from protobuf enum value: None = 0;
     */
    None = 0,
    /**
     * @generated from protobuf enum value: Name = 1;
     */
    Name = 1,
    /**
     * @generated from protobuf enum value: LastRead = 2;
     */
    LastRead = 2,
    /**
     * @generated from protobuf enum value: LatestRelease = 3;
     */
    LatestRelease = 3
}
/**
 * @generated from protobuf enum wuxiaworld.api.v2.GetBookmarksRequest.SortDirection
 */
export enum GetBookmarksRequest_SortDirection {
    /**
     * @generated from protobuf enum value: ASC = 0;
     */
    ASC = 0,
    /**
     * @generated from protobuf enum value: DESC = 1;
     */
    DESC = 1
}
// ...
export enum ReadingListType {
    /**
     * @generated from protobuf enum value: Current = 0;
     */
    Current = 0,
    /**
     * @generated from protobuf enum value: Completed = 1;
     */
    Completed = 1,
    /**
     * @generated from protobuf enum value: Dropped = 3;
     */
    Dropped = 3,
    /**
     * @generated from protobuf enum value: OnHold = 4;
     */
    OnHold = 4,
    /**
     * @generated from protobuf enum value: Manual = 5;
     */
    Manual = 5
}
// @generated message type with reflection information, may provide speed optimized methods
class GetBookmarksRequest$Type extends MessageType<GetBookmarksRequest> {
    constructor() {
        super("wuxiaworld.api.v2.GetBookmarksRequest", [
            { no: 1, name: "listType", kind: "enum", T: () => ["wuxiaworld.api.v2.ReadingListType", ReadingListType] },
            { no: 2, name: "novelId", kind: "scalar", oneof: "selector", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "novelSlug", kind: "scalar", oneof: "selector", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "pageInfo", kind: "message", oneof: "pagination", T: () => PageInfoRequest },
            { no: 5, name: "cursor", kind: "message", oneof: "pagination", T: () => GetBookmarksRequest_CursorRequest },
            { no: 6, name: "sortBy", kind: "enum", T: () => ["wuxiaworld.api.v2.GetBookmarksRequest.SortBy", GetBookmarksRequest_SortBy] },
            { no: 7, name: "sortDirection", kind: "enum", T: () => ["wuxiaworld.api.v2.GetBookmarksRequest.SortDirection", GetBookmarksRequest_SortDirection] }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.GetBookmarksRequest
 */
export const GetBookmarksRequest = new GetBookmarksRequest$Type();
// @generated message type with reflection information, may provide speed optimized methods
class GetBookmarksRequest_CursorRequest$Type extends MessageType<GetBookmarksRequest_CursorRequest> {
    constructor() {
        super("wuxiaworld.api.v2.GetBookmarksRequest.CursorRequest", [
            { no: 1, name: "chapterOffset", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "count", kind: "scalar", T: 5 /*ScalarType.INT32*/, options: { "wuxiaworld.api.v2.range": { max: { value: 100 } } } }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message wuxiaworld.api.v2.GetBookmarksRequest.CursorRequest
 */
export const GetBookmarksRequest_CursorRequest = new GetBookmarksRequest_CursorRequest$Type();